requirements.txt
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
